+++ .github/workflows/test.yml
@@ -1,0 +1,39 @@
+name: Run Tests
+
+on:
+  push:
+    branches:
+      - main
+
+  pull_request:
+
+jobs:
+  # Run tests
+  test:
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v2
+
+      - name: Log in to package registry
+        uses: docker/login-action@v1
+        with:
+          registry: ghcr.io
+          username: ${{ github.actor }}
+          password: ${{ github.token }}
+
+      - name: Set up tests
+        run: |
+          docker-compose build
+          docker-compose run web bundle install
+
+      - name: Run standardrb
+        run: docker-compose run web bundle exec standardrb
+
+      - name: Run tests
+        run: docker-compose run test
+
+      - name: Report to Coveralls
+        uses: coverallsapp/github-action@1.1.3
+        with:
+          github-token: ${{ secrets.github_token }}
