+++ Gemfile
@@ -33,4 +33,6 @@ gem 'pry'
 gem 'simple_solr_client'
 
 gem 'zinzout'
-gem 'thor'
+gem "thor"
+gem "dotenv"
+gem "yell"
