+++ docker-compose.yml
@@ -25,7 +25,10 @@ services:
       - gem_cache:/gems
     command: bash -c "wait-for solr-sdr-catalog:9033 && bundle exec rspec"
     environment:
-      SOLR_URL: http://solr-sdr-catalog:9033/solr/catalog
+      - SOLR_URL=http://solr-sdr-catalog:9033/solr/catalog
+      - redirect_file=/dev/null
+      - NO_DB=1
+      - DDIR=/app/examples
     depends_on:
       - solr-sdr-catalog
